+++ packages/cds-odata-annotation-converter/src/transforms/annotation/visitor-state.ts
@@ -89,6 +89,7 @@ export class VisitorState {
     pushContext(context: Context) {
         this._contextStack.push(Object.seal({ ...context }));
     }
+
     popContext() {
         this._contextStack.splice(-1);
     }