+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/fpm_v4/webapp/test/integration/pages/TravelMain.js
@@ -1,0 +1,13 @@
+sap.ui.define(['sap/fe/test/TemplatePage'], function(TemplatePage) {
+    'use strict';
+
+    var CustomPageDefinitions = {
+        actions: {},
+        assertions: {}
+    };
+
+    return new TemplatePage(
+        'testNameSpace.fpmv4::TravelMain',
+        CustomPageDefinitions
+    );
+});