+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/headless/lrop_v4_no_ui5_version/webapp/test/integration/pages/TravelList.js
@@ -1,0 +1,17 @@
+sap.ui.define(['sap/fe/test/ListReport'], function(ListReport) {
+    'use strict';
+
+    var CustomPageDefinitions = {
+        actions: {},
+        assertions: {}
+    };
+
+    return new ListReport(
+        {
+            appId: 'testnamepsace.lropv4noui5version',
+            componentId: 'TravelList',
+            contextPath: '/Travel'
+        },
+        CustomPageDefinitions
+    );
+});