+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/cap-config-writer
 
+## 0.9.25
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.10.2
+
+## 0.9.24
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.10.1
+
+## 0.9.23
+
+### Patch Changes
+
+-   23e055a: Adds new module @sap-ux/fiori-app-sub-generator
+-   Updated dependencies [23e055a]
+    -   @sap-ux/fiori-generator-shared@0.10.0
+
 ## 0.9.22
 
 ### Patch Changes