+++ packages/launch-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/launch-config
 
+## 0.9.0
+
+### Minor Changes
+
+-   b49f8d5: Export handleWorkspaceConfig from launch-config and use in adt app download sub generator
+
+## 0.8.1
+
+### Patch Changes
+
+-   23e055a: Adds new module @sap-ux/fiori-app-sub-generator
+
 ## 0.8.0
 
 ### Minor Changes